+++ src/redux/actionTypes/actionTypes.js
@@ -1,0 +1,9 @@
+// Actions
+const ADD_BOOK = 'bookstore/books/ADD_BOOK';
+const REMOVE_BOOK = 'bookstore/books/REMOVE_BOOK';
+const RETRIEVE_BOOK = 'bookstore/books/RETRIEVE_BOOK';
+const STATUS = 'bookstore/categories/status';
+
+export {
+  ADD_BOOK, REMOVE_BOOK, RETRIEVE_BOOK, STATUS,
+};
