+++ src/redux/actions/categories.js
@@ -1,0 +1,6 @@
+import { STATUS } from '../actionTypes/actionTypes';
+
+// Action Creators
+const checkStatus = () => ({ type: STATUS });
+
+export default checkStatus;
