+++ src/redux/store/configureStore.js
@@ -2,8 +2,8 @@ import { createStore, combineReducers, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
-import bookReducer from './books/books';
-import categoryReducer from './categories/categories';
+import bookReducer from '../reducers/books';
+import categoryReducer from '../reducers/categories';
 
 const rootReducer = combineReducers({ bookReducer, categoryReducer });
 const middleware = [logger, thunk];
